modularizing data, and decoupling data. still need to decouple some more but these are working changes.
